feat: Client-side bookmark parsing for fast edit dialogs
- Added `bookmarks_parser.gohtml` with JavaScript parsing functions (`extractTabByIndex`, `extractPage`, `extractCategoryByIndex`) to extract bookmark blocks on the client side.
- Updated `tail.gohtml` to intercept edit link clicks and populate the modal client-side using the new parser, removing the need for a server roundtrip.
- Embedded raw bookmarks into a hidden `<textarea>` to allow the browser to natively handle HTML unescaping.
- Added client-side `<template>` blocks for tab, page, and category edit forms including required `branch`, `ref`, and `sha` inputs.
- Created `tests_js/parse_test.js` to verify the logic directly with Node.js instead of Go.
- Registered the 'bookmarks' func map in test helpers (`data_test.go`) to avoid test failures.
- Cleaned up backup files resulting from patching operations.
